Add delete functionality
I tried to get delete to work too, and it does so long as you refresh the
page after hitting the delete button. I ran out of time to get the page to
refresh after adding or deleting without you having to manually refresh the
page. Styling wasn't added in time.